@@ -10,19 +10,33 @@
     "ionic:serve": "ionic-app-scripts serve"
   },
   "dependencies": {
-    "@angular/common": "4.0.2",
-    "@angular/compiler": "4.0.2",
-    "@angular/compiler-cli": "4.0.2",
-    "@angular/core": "4.0.2",
-    "@angular/forms": "4.0.2",
-    "@angular/http": "4.0.2",
-    "@angular/platform-browser": "4.0.2",
-    "@angular/platform-browser-dynamic": "4.0.2",
+    "@angular/common": "4.3.3",
+    "@angular/compiler": "4.3.3",
+    "@angular/compiler-cli": "4.3.3",
+    "@angular/core": "4.3.3",
+    "@angular/forms": "4.3.3",
+    "@angular/http": "4.3.3",
+    "@angular/platform-browser": "4.3.3",
+    "@angular/platform-browser-dynamic": "4.3.3",
     "@ionic-native/core": "3.6.1",
+    "@ionic-native/estimote-beacons": "^4.20.0",
+    "@ionic-native/ibeacon": "^4.20.0",
+    "@ionic-native/in-app-browser": "^4.20.0",
     "@ionic-native/splash-screen": "3.6.1",
     "@ionic-native/status-bar": "3.6.1",
     "@ionic/storage": "2.0.1",
+    "cordova-android": "7.1.4",
+    "cordova-plugin-ble": "^2.0.1",
+    "cordova-plugin-console": "^1.1.0",
+    "cordova-plugin-device": "^1.1.7",
+    "cordova-plugin-eddystone": "^1.3.0",
+    "cordova-plugin-estimote": "0.8.0",
+    "cordova-plugin-ibeacon": "3.8.1",
+    "cordova-plugin-splashscreen": "^4.1.0",
+    "cordova-plugin-statusbar": "^2.4.3",
+    "cordova-plugin-whitelist": "^1.3.4",
     "ionic-angular": "3.1.1",
+    "ionic-plugin-keyboard": "^2.2.1",
     "ionicons": "3.0.0",
     "rxjs": "5.1.1",
     "sw-toolbox": "3.4.0",
@@ -48,5 +62,21 @@
       "locator": "ios"
     }
   ],
-  "description": "scrolltabs: An Ionic project"
+  "description": "scrolltabs: An Ionic project",
+  "cordova": {
+    "plugins": {
+      "ionic-plugin-keyboard": {},
+      "cordova-plugin-whitelist": {},
+      "cordova-plugin-console": {},
+      "cordova-plugin-statusbar": {},
+      "cordova-plugin-device": {},
+      "cordova-plugin-splashscreen": {},
+      "cordova-plugin-estimote": {},
+      "com.unarin.cordova.beacon": {},
+      "cordova-plugin-eddystone": {}
+    },
+    "platforms": [
+      "android"
+    ]
+  }
 }
